leader restructure
